add notes and change max and min

+++ Board.hpp
@@ -1,5 +1,6 @@
 #include <string>
 #include "Direction.hpp"
+#include <climits>
 #include <map>
 
 using namespace std;
@@ -8,9 +9,9 @@ namespace ariel{
     class Board{
     private:
         unsigned int maxrow=0;
-        unsigned int minrow=INT8_MAX;
+        unsigned int minrow=UINT_MAX;
         unsigned int maxcols=0;
-        unsigned int mincols=INT8_MAX;
+        unsigned int mincols=UINT_MAX;
         std::map<unsigned int,std::map<unsigned int,string>> board;
     public:
         void post(unsigned int row, unsigned int column, Direction dierction, string message);
